Create a JSON Songs File in Python and Code Beautify
Create a JSON Songs File

View of a sample JSON data set.  Look for the structure and the notion of key / value pairs.  It is good to practice with the JSON format, because it is very similar to a dictionary data type.  It is good to understand the concept of key / value pairs, because they are used frequently in the IT field.  This also further reiterates the usage of different data types, such as strings and integers.

How to do this assignment, which includes:
How to download a JSON file
How to do this assignment, using the PyCharm IDE
How to do this assignment using the IDLE IDE

(You can use any IDE or text editor to do this assignment.  JSON is text-based, which means that you can use any text editor to update it, it is simply easier to view with a "smart" editor, such as PyCharm, because it will show the colors.)

Use CodeBeautify to error check:
https://codebeautify.org/jsonviewer